move type defenitions
