+++ front/types/translation.ts
